requirements.txt
asgiref==3.8.1
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
sqlparse==0.5.1
